# api/openapi.yaml
openapi: 3.1.0
info:
  title: Order Management System API
  version: 0.2.0
  description: V1 REST contract for the Order Management System. This contract covers only health checks and order lifecycle APIs within the approved V1 OMS scope. Provider-neutral Bearer JWT verification is available with `OMS_AUTH_MODE=jwt` and configured OIDC discovery or JWKS metadata. Local development may explicitly enable trusted headers with `OMS_AUTH_MODE=dev`; those headers are not production authentication.
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local development server
tags:
  - name: system
    description: Service health endpoints
  - name: orders
    description: Customer and admin order operations
paths:
  /healthz:
    get:
      security: []
      tags:
        - system
      summary: Health check
      operationId: getHealthz
      responses:
        "200":
          description: Server is healthy.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
  /api/v1/orders:
    post:
      tags:
        - orders
      summary: Create order
      description: Creates a new order for the authenticated customer. The customer identity comes from the bearer token, and the server computes `total_cents`.
      operationId: createOrder
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: Idempotency conflict or other order state conflict.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
    get:
      tags:
        - orders
      summary: List owned orders
      description: Lists orders owned by the authenticated customer using keyset pagination.
      operationId: listOrders
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          description: Optional order status filter.
          required: false
          schema:
            $ref: "#/components/schemas/OrderStatus"
        - name: cursor
          in: query
          description: Opaque keyset pagination cursor returned from a previous list response.
          required: false
          schema:
            type: string
            minLength: 1
        - name: limit
          in: query
          description: Maximum number of orders to return in a single page.
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: A page of orders.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
        "400":
          description: Invalid status filter, malformed cursor, or invalid limit.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/orders/{order_id}:
    get:
      tags:
        - orders
      summary: Retrieve owned order by ID
      description: Returns an order only when it belongs to the authenticated customer.
      operationId: getOrderByID
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OrderID"
      responses:
        "200":
          description: The requested order.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: The caller is authenticated but not allowed to access this order.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Order not found or not owned by the authenticated customer.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/orders/{order_id}/cancel:
    post:
      tags:
        - orders
      summary: Cancel owned order
      description: Cancels an order only when it belongs to the authenticated customer and is currently `PENDING`.
      operationId: cancelOrder
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OrderID"
      responses:
        "200":
          description: Updated order after cancellation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: The caller is authenticated but not allowed to cancel this order.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Order not found or not owned by the authenticated customer.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: The order cannot be cancelled because it is not in `PENDING` status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/orders/{order_id}/status:
    patch:
      tags:
        - orders
      summary: Update order status
      description: Admin/system-only endpoint for moving an order to a new status.
      operationId: updateOrderStatus
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OrderID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusRequest"
      responses:
        "200":
          description: Updated order.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Missing or invalid status value.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: The authenticated caller is not an admin or system actor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Order not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Invalid status transition.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    OrderID:
      name: order_id
      in: path
      required: true
      description: Order identifier.
      schema:
        type: string
        minLength: 1
  responses:
    BadRequest:
      description: Invalid request input.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalServerError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    HealthResponse:
      type: object
      additionalProperties: false
      required:
        - status
      properties:
        status:
          type: string
          const: ok
    OrderStatus:
      type: string
      enum:
        - PENDING
        - PROCESSING
        - SHIPPED
        - DELIVERED
        - CANCELLED
    CurrencyCode:
      type: string
      pattern: "^[A-Z]{3}$"
      description: ISO 4217 currency code.
    OrderItemInput:
      type: object
      additionalProperties: false
      required:
        - product_id
        - sku
        - quantity
        - unit_price_cents
      properties:
        product_id:
          type: string
          minLength: 1
        sku:
          type: string
          minLength: 1
        quantity:
          type: integer
          minimum: 1
        unit_price_cents:
          type: integer
          minimum: 0
    CreateOrderRequest:
      type: object
      additionalProperties: false
      required:
        - idempotency_key
        - currency
        - items
      properties:
        idempotency_key:
          type: string
          minLength: 1
        currency:
          $ref: "#/components/schemas/CurrencyCode"
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/OrderItemInput"
    OrderItem:
      type: object
      additionalProperties: false
      required:
        - product_id
        - sku
        - quantity
        - unit_price_cents
      properties:
        product_id:
          type: string
        sku:
          type: string
        quantity:
          type: integer
        unit_price_cents:
          type: integer
    Order:
      type: object
      additionalProperties: false
      required:
        - id
        - customer_id
        - status
        - currency
        - total_cents
        - items
        - created_at
        - updated_at
      properties:
        id:
          type: string
          minLength: 1
        customer_id:
          type: string
          minLength: 1
          description: Authenticated customer identifier associated with the order.
        status:
          $ref: "#/components/schemas/OrderStatus"
        currency:
          $ref: "#/components/schemas/CurrencyCode"
        total_cents:
          type: integer
          minimum: 0
          description: Computed server-side from the order items.
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/OrderItem"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    OrderListResponse:
      type: object
      additionalProperties: false
      required:
        - orders
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        next_cursor:
          type:
            - string
            - "null"
          description: Opaque keyset cursor for the next page. `null` means there is no subsequent page.
    UpdateOrderStatusRequest:
      type: object
      additionalProperties: false
      required:
        - status
      properties:
        status:
          $ref: "#/components/schemas/OrderStatus"
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          type: object
          additionalProperties: false
          required:
            - code
            - message
          properties:
            code:
              type: string
              minLength: 1
            message:
              type: string
              minLength: 1