# internal/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package db

import (
	"net/netip"

	"github.com/jackc/pgx/v5/pgtype"
)

type Order struct {
	ID             pgtype.UUID
	CustomerID     pgtype.UUID
	Status         string
	IdempotencyKey string
	TotalCents     int64
	Currency       string
	CreatedAt      pgtype.Timestamptz
	UpdatedAt      pgtype.Timestamptz
}

type OrderEvent struct {
	ID         pgtype.UUID
	OrderID    pgtype.UUID
	ActorType  string
	ActorID    pgtype.UUID
	FromStatus pgtype.Text
	ToStatus   pgtype.Text
	Action     string
	Reason     pgtype.Text
	IpAddress  *netip.Addr
	CreatedAt  pgtype.Timestamptz
}

type OrderItem struct {
	ID             pgtype.UUID
	OrderID        pgtype.UUID
	ProductID      pgtype.UUID
	Sku            string
	Quantity       int32
	UnitPriceCents int64
	CreatedAt      pgtype.Timestamptz
}